% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Displays the central limit for a poisson distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Size of sample}

\item{iter}{Number of samples}

\item{lambda}{The lambda paramater for the poisson distribution}

\item{...}{Extra paramaters for histogram display}
}
\value{
Graphs of the means of the samples and the distribution compared to the theoretical poisson
}
\description{
Displays the central limit for a poisson distribution
}
\examples{
mycltp(n=10,iter=10000,lambda=2)
mycltp(4,20)
}
