% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiform.R
\name{runiform}
\alias{runiform}
\title{Randomly generates a uniform distribution and displays it}
\usage{
runiform(k, n, iter = 10, time = 0.5)
}
\arguments{
\item{k}{Size of distribution}

\item{n}{Number of iterations}

\item{iter}{Number of iterations to display}

\item{time}{Time between graph displays}
}
\value{
Graph of uniform distribution
}
\description{
Randomly generates a uniform distribution and displays it
}
\examples{
runiform(10,100)
runiform(10,100, 5,0.2)
}
