% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sos.R
\name{sumofsquares}
\alias{sumofsquares}
\title{Display sum of squares info}
\usage{
sumofsquares(linearmodel, datavariable)
}
\arguments{
\item{linearmodel}{A linear model.}

\item{datavariable}{A variable.}
}
\value{
Sum of squares info about the linear model.
}
\description{
Display sum of squares info
}
\examples{
sumofsquares(spruce.lm, spruce.df$Height)
sumofsquares(quad.lm, spruce.df$Height)
}
